updated fps

+++ src/scripts/VideoQueue.py
@@ -8,7 +8,7 @@ import os
 from sensor_msgs.msg import Image
 from std_srvs.srv import Empty, EmptyResponse  # Import the Empty service type and its response
 from time import time
-
+time=time
 class VideoQueue(object):
     def __init__(self):
         self.bridge = CvBridge()
@@ -16,7 +16,7 @@ class VideoQueue(object):
         self.depth_sub = rospy.Subscriber('/camera/aligned_depth_to_color/image_raw', Image, self.depth_callback)
         self.pause_service = rospy.Service('pause_video_capture', Empty, self.pause_callback)
         self.resume_service = rospy.Service('resume_video_capture', Empty, self.resume_callback)  # Add a service to resume video capture
-        self.window = 5
+        self.window = 3
         self.pause = False  # Initialize the pause attribute
         # Create directories if they don't exist
         self.color_dir = os.path.expanduser('~/catkin_ws/src/guide_dog/src/tmp/color')
@@ -31,23 +31,26 @@ class VideoQueue(object):
         if self.pause:
             return
 
-        rate = rospy.Rate(1)  # 1 fps
+        rate = rospy.Rate(5)  # 1 fps
 
         number_of_files = len(os.listdir(self.depth_dir))
         if number_of_files < self.window:
             cv2.imwrite(f'{self.color_dir}/{number_of_files}.png', self.color_image)
-            self.depth_image = self.bridge.imgmsg_to_cv2(data, desired_encoding="16UC1")
+            self.depth_image = self.bridge.imgmsg_to_cv2(data, desired_encoding='passthrough')
             np.save(f'{self.depth_dir}/{number_of_files}.npy', self.depth_image)
             print(f"Saved image {number_of_files}.npy' at {time()}")
             rate.sleep()
             return
 
-        self.depth_image = self.bridge.imgmsg_to_cv2(data, desired_encoding="16UC1")
-        os.remove(f'{self.color_dir}/0.png')
-        os.remove(f'{self.depth_dir}/0.npy')
-        for j in range(1, self.window):
-            os.rename(f'{self.color_dir}/{j}.png', f'{self.color_dir}/{j-1}.png')
-            os.rename(f'{self.depth_dir}/{j}.npy', f'{self.depth_dir}/{j-1}.npy')
+        self.depth_image = self.bridge.imgmsg_to_cv2(data, desired_encoding="passthrough")
+        self.depth_image = np.clip(self.depth_image, None, 10000)
+
+        if os.path.exists(f'{self.color_dir}/0.png'):
+            os.remove(f'{self.color_dir}/0.png')
+            os.remove(f'{self.depth_dir}/0.npy')
+            for j in range(1, self.window):
+                os.rename(f'{self.color_dir}/{j}.png', f'{self.color_dir}/{j-1}.png')
+                os.rename(f'{self.depth_dir}/{j}.npy', f'{self.depth_dir}/{j-1}.npy')
 
         cv2.imwrite(f'{self.color_dir}/{self.window-1}.png', self.color_image)
         np.save(f'{self.depth_dir}/{self.window-1}.npy', self.depth_image)
